Flytt valideringslogikk inn i egen component
